Add files via upload
In my capstone project, I created a web-based token staking application using React and Ethereum blockchain. It allows users to deposit, withdraw, trade, and join in airdrop events. I'm proud of this user-friendly and secure application and believe it will help me in my future job as a Blockchain developer.